Hack back-branch SSL tests to avoid intermittent buildfarm failures.
Buildfarm member eelpout sometimes reports the wrong error message for
an SSL connection failure.  In HEAD, this problem is believed to be
solved by commit 1f39a1c06, but I'm as yet unwilling to back-patch that.
The problem seems fairly unlikely to be an issue in the field, since (as
far as we can tell) it happens only during a failure of a local-loopback
SSL connection, and it's improbable even then.  It seems better to just
live with it for the time being; but let's tweak the regression test to
accept the other error message as a "pass".

Needed in v11 only, since older branches didn't check the message
text anyway.

Discussion: https://postgr.es/m/CAEepm=2n6Nv+5tFfe8YnkUm1fXgvxR0Mm1FoD+QKG-vLNGLyKg@mail.gmail.com